android/pyproject.toml
[project]
name = "remotivelabs-topology-android-example"
description = "RemotiveTopology android example"
version = "1.0.0"
readme = "README.md"
license-files = ["LICENSE"]

requires-python = ">=3.9,<4.0"

dependencies = [
  "structlog==25.4.0",
  "pytest==8.3.5",
  "pytest-asyncio==1.0.0",
  "transitions==0.9.2",
  "jupyterlab >=4.3.4",
  "ipywidgets >=8.1.5",
  "pillow >=11.1.0",
  "gherkin-official >=30.0.4",
  "behave >=1.2.6",
  "remotivelabs-broker~=0.9.0b3",
  "remotivelabs-topology~=0.9.0",
]

[tool.setuptools.packages.find]
where = ["ecus/"]
include = ["bcm*", "gwm*", "ihu*"]
namespaces = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=80.0.0", "wheel"]
build-backend = "setuptools.build_meta"
